pyproject.toml
[tool.poetry]
name = "python-server-adapter"
version = "0.1.0"
description = ""
authors = ["mr-fox-97 <eric.m.cardozo@mi.unc.edu.ar>"]
license = "MIT LICENSE"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.31"}
pytest-asyncio = "^0.23.8"
aioredis = "^2.0.1"
python-dotenv = "^1.0.1"
pydantic = "^2.8.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
